+++ htdocs/js/Search.js
@@ -116,20 +116,23 @@ function initSearch() {
             event.stopPropagation();
             event.target.parentNode.remove();
             doSearch('');
+            document.getElementById('searchStr').updateBtn();
         }
         else if (event.target.nodeName === 'BUTTON') {
             //edit search expression
             event.preventDefault();
             event.stopPropagation();
-            document.getElementById('searchStr').value = unescapeMPD(getData(event.target, 'filter-value'));
+            const searchStrEl = document.getElementById('searchStr');
+            searchStrEl.value = unescapeMPD(getData(event.target, 'filter-value'));
             selectTag('searchTags', 'searchTagsDesc', getData(event.target, 'filter-tag'));
             document.getElementById('searchMatch').value = getData(event.target, 'filter-op');
             event.target.remove();
             app.current.filter = getData(event.target,'filter-tag');
-            doSearch(document.getElementById('searchStr').value);
+            doSearch(searchStrEl.value);
             if (document.getElementById('searchCrumb').childElementCount === 0) {
                 elHideId('searchCrumb');
             }
+            searchStrEl.updateBtn();
         }
     }, false);