+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This is a small bugfix release.
 - Feat: add pl-PL translation
 - Upd: remove zh-Hant translation
 - Fix: JS error when clearing the search box #982
+- Fix: update clear button for inputs on breadcrumb select
 
 ***